+++ Realm/Tests/Swift/SwiftRealmTests.swift
@@ -47,22 +47,22 @@ class SwiftRealmTests: SwiftTestCase {
         SwiftStringObject.createInRealm(realm, withObject: ["a"])
         SwiftStringObject.createInRealm(realm, withObject: ["b"])
         SwiftStringObject.createInRealm(realm, withObject: ["c"])
-        XCTAssertEqual(SwiftStringObject.allObjectsInRealm(realm).count, 3, "Expecting 3 objects")
+        XCTAssertEqual(SwiftStringObject.allObjectsInRealm(realm).count, UInt(3), "Expecting 3 objects")
         realm.commitWriteTransaction()
 
         // test again after write transaction
         var objects = SwiftStringObject.allObjectsInRealm(realm)
-        XCTAssertEqual(objects.count, 3, "Expecting 3 objects")
+        XCTAssertEqual(objects.count, UInt(3), "Expecting 3 objects")
         XCTAssertEqual((objects[0] as SwiftStringObject).stringCol, "a", "Expecting column to be 'a'")
 
         realm.beginWriteTransaction()
         realm.deleteObject(objects[2] as SwiftStringObject)
         realm.deleteObject(objects[0] as SwiftStringObject)
-        XCTAssertEqual(SwiftStringObject.allObjectsInRealm(realm).count, 1, "Expecting 1 object")
+        XCTAssertEqual(SwiftStringObject.allObjectsInRealm(realm).count, UInt(1), "Expecting 1 object")
         realm.commitWriteTransaction()
 
         objects = SwiftStringObject.allObjectsInRealm(realm)
-        XCTAssertEqual(objects.count, 1, "Expecting 1 object")
+        XCTAssertEqual(objects.count, UInt(1), "Expecting 1 object")
         XCTAssertEqual((objects[0] as SwiftStringObject).stringCol, "b", "Expecting column to be 'b'")
     }
 
@@ -87,7 +87,7 @@ class SwiftRealmTests: SwiftTestCase {
 
         // get object
         let objects = SwiftStringObject.allObjectsInRealm(realm)
-        XCTAssertEqual(objects.count, 1, "There should be 1 object of type StringObject")
+        XCTAssertEqual(objects.count, UInt(1), "There should be 1 object of type StringObject")
         XCTAssertEqual((objects[0] as SwiftStringObject).stringCol, "string", "Value of first column should be 'string'")
     }
 
@@ -107,7 +107,7 @@ class SwiftRealmTests: SwiftTestCase {
         realm.commitWriteTransaction()
 
         let objects = SwiftIgnoredPropertiesObject.allObjectsInRealm(realm)
-        XCTAssertEqual(objects.count, 1, "There should be 1 object of type SwiftIgnoredPropertiesObject")
+        XCTAssertEqual(objects.count, UInt(1), "There should be 1 object of type SwiftIgnoredPropertiesObject")
         XCTAssertNil((objects[0] as SwiftIgnoredPropertiesObject).runtimeProperty, "Ignored property should be nil")
         XCTAssertEqual((objects[0] as SwiftIgnoredPropertiesObject).name, "@fz", "Value of the name column doesn't match the assigned one.")
     }
@@ -119,9 +119,9 @@ class SwiftRealmTests: SwiftTestCase {
         let updateComplete = expectationWithDescription("background update complete")
 
         let token = realm.addNotificationBlock() { (_, _) in
-            XCTAssertEqual(objs.count, 2)
-            XCTAssertEqual(objs.sortedResultsUsingProperty("intCol", ascending: true).count, 2)
-            XCTAssertEqual(objects.count, 2)
+            XCTAssertEqual(objs.count, UInt(2))
+            XCTAssertEqual(objs.sortedResultsUsingProperty("intCol", ascending: true).count, UInt(2))
+            XCTAssertEqual(objects.count, UInt(2))
             updateComplete.fulfill()
         }
 
@@ -159,7 +159,7 @@ class SwiftRealmTests: SwiftTestCase {
             realm.commitWriteTransaction()
 
             let objects = SwiftStringObject.allObjectsInRealm(realm)
-            XCTAssertEqual(objects.count, 1, "There should be 1 object of type StringObject")
+            XCTAssertEqual(objects.count, UInt(1), "There should be 1 object of type StringObject")
             XCTAssertEqual((objects[0] as SwiftStringObject).stringCol, "string", "Value of first column should be 'string'")
         }
 
@@ -168,7 +168,7 @@ class SwiftRealmTests: SwiftTestCase {
 
         // get object
         let objects = SwiftStringObject.allObjectsInRealm(realm)
-        XCTAssertEqual(objects.count, 1, "There should be 1 object of type RLMTestObject")
+        XCTAssertEqual(objects.count, UInt(1), "There should be 1 object of type RLMTestObject")
         XCTAssertEqual((objects[0] as SwiftStringObject).stringCol, "string", "Value of first column should be 'string'")
     }
 
@@ -180,22 +180,22 @@ class SwiftRealmTests: SwiftTestCase {
         StringObject.createInRealm(realm, withObject: ["a"])
         StringObject.createInRealm(realm, withObject: ["b"])
         StringObject.createInRealm(realm, withObject: ["c"])
-        XCTAssertEqual(StringObject.allObjectsInRealm(realm).count, 3, "Expecting 3 objects")
+        XCTAssertEqual(StringObject.allObjectsInRealm(realm).count, UInt(3), "Expecting 3 objects")
         realm.commitWriteTransaction()
 
         // test again after write transaction
         var objects = StringObject.allObjectsInRealm(realm)
-        XCTAssertEqual(objects.count, 3, "Expecting 3 objects")
+        XCTAssertEqual(objects.count, UInt(3), "Expecting 3 objects")
         XCTAssertEqual((objects[0] as StringObject).stringCol!, "a", "Expecting column to be 'a'")
 
         realm.beginWriteTransaction()
         realm.deleteObject(objects[2] as StringObject)
         realm.deleteObject(objects[0] as StringObject)
-        XCTAssertEqual(StringObject.allObjectsInRealm(realm).count, 1, "Expecting 1 object")
+        XCTAssertEqual(StringObject.allObjectsInRealm(realm).count, UInt(1), "Expecting 1 object")
         realm.commitWriteTransaction()
 
         objects = StringObject.allObjectsInRealm(realm)
-        XCTAssertEqual(objects.count, 1, "Expecting 1 object")
+        XCTAssertEqual(objects.count, UInt(1), "Expecting 1 object")
         XCTAssertEqual((objects[0] as StringObject).stringCol!, "b", "Expecting column to be 'b'")
     }
 
@@ -222,7 +222,7 @@ class SwiftRealmTests: SwiftTestCase {
 
         // get object
         let objects = StringObject.allObjectsInRealm(realm)
-        XCTAssertEqual(objects.count, 1, "There should be 1 object of type StringObject")
+        XCTAssertEqual(objects.count, UInt(1), "There should be 1 object of type StringObject")
         XCTAssertEqual((objects[0] as StringObject).stringCol!, "string", "Value of first column should be 'string'")
     }
 
@@ -244,7 +244,7 @@ class SwiftRealmTests: SwiftTestCase {
             }
 
             let objects = StringObject.allObjectsInRealm(realm)
-            XCTAssertEqual(objects.count, 1, "There should be 1 object of type StringObject")
+            XCTAssertEqual(objects.count, UInt(1), "There should be 1 object of type StringObject")
             XCTAssertEqual((objects[0] as StringObject).stringCol!, "string", "Value of first column should be 'string'")
         }
 
@@ -253,7 +253,7 @@ class SwiftRealmTests: SwiftTestCase {
 
         // get object
         let objects = StringObject.allObjectsInRealm(realm)
-        XCTAssertEqual(objects.count, 1, "There should be 1 object of type RLMTestObject")
+        XCTAssertEqual(objects.count, UInt(1), "There should be 1 object of type RLMTestObject")
         XCTAssertEqual((objects[0] as StringObject).stringCol!, "string", "Value of first column should be 'string'")
     }
 }
